requirements.txt
streamlit
pandas
numpy
matplotlib
scipy
google-generativeai